Adding "Match" (with tests).
This is a logical representation of a match between
2 players.
It keeps track of the number of players, maintaining a useful
status to know at which "point" in the game we are.

It will be useful to represent the concept of "Room" where max 2 players are allowed
to join and play.